chore(fluxWizard): refactor costants
