+++ ui/src/shared/components/TimeMachine/fluxQueryBuilder/util/constants.ts
